Add new `Arrow.tsx` file to commonComponents
A new `Arrow.tsx` file has been added to the commonComponents folder. It defines and exports a motion-enabled arrow component using Framer Motion. This new component will be used to improve the design and interactivity of several UI elements. The paths and gradients of the arrow have been clearly defined in this file.