Refactor Arrow component, improve its style and animation
This commit introduces a significant refactor of the Arrow component. Originally, MotionImage was used to display the arrow but was changed to MotionSvg for better customizability. The Arrow component was moved to the commonComponents directory, making it more accessible across our application. Also, changes were made to the color property of the SectionWrapper component from "common.white" to "#e6e6e6" to better match our design aesthetic.

Importantly, the refactor introduces animations to the Arrow component, enhancing the visual appeal and user interaction. This change is driven by a desire to improve UI/UX. It should provide a better experience for the users, allowing them to interact with the app more fluidly.