Update next.js and Swiper versions
The project's dependencies have been updated, particularly the versions of next.js and Swiper have been bumped up to the latest versions - 13.5.2 and 10.3.0, respectively. These updates ensure we get the latest fixes, enhancements, and improvements provided by these libraries. Regular updating of dependencies is crucial in maintaining the robustness and performance of the application. The integrity and resolved URLs in package-lock.json have also been updated in alignment with these changes.