Update React Fragment in home/impact.tsx
Switched from using shorthand( <></> ) to explicit syntax( <React.Fragment></React.Fragment> ) for the React Fragment in the home/impact.tsx file. The change also includes adding a unique 'key' prop to each child in a list. This helps React identify which items have changed, are added, or are removed and helps in efficient re-rendering of the components.