Update responsive breakpoints and styles for site components
Made changes to the responsive breakpoints and associated styles in multiple files to ensure a smoother transition between screen sizes. Updated the media queries and values such as padding and margin in components like NavigationActions and SectionWrapper. Adjusted the breakpoints in the theme file. Changed the display property for navigation in desktop mode from 'xl' to 'lg_xl' to make it more responsive. Moved appbarHeight from the header to reactive-appbar component for a cleaner structure. This will improve the UI/UX for screens of various sizes.