Change color and refactor code in nav components
Changed the color of the ButtonGroup from secondary to primary in the desktop-navigation.tsx. Refactored the NavItemProps interface in the nav-item.tsx for a more simplified extension. Removed 'MLinkProps' and used 'restButtonProps' for more concise usage.

These changes are made for better color contrast in header navigation and cleaner, more maintainable codebase by reducing unnecessary use of props.