Adjust typography settings and improve responsive design
In response to feedback and tests, some tweaks to the typography settings have been made. The margin-bottom property (mb) of gutterBottom in typography.theme.ts was changed to better space out paragraphs, especially on larger screens. Similarly, changes in in-the-news.tsx and arrow-slide-info.tsx were made to better adapt the sections for different screen sizes. This adjustment improves the responsiveness of the website, ensuring better view on various devices and screen sizes. These changes contribute to better readability, more balanced layout, and overall enhanced user interface.