Refactor navigation settings and improve swiper instance
In the `arrow-swiper.tsx` file, the navigation settings have been refactored to simplify the use of the `navigation` property. Instead of using a conditional ternary operator, a spread operator is used to include `navigation` in the general settings. This should make the code cleaner and more maintainable. The `InTheNews` component was also updated to use the `useRef` hook to access the swiper instance. This should prevent unnecessary re-renders and improve performance.