Refactor files to improve navigation and animations
The files arrow-swiper-with-info-section.tsx and arrow-slide-default-image.tsx were modified to improve the interaction and functionality of the components. The title property in ArrowSlideDefaultImageProps was updated to make it optional to increase the component's flexibility. Also, The 'SwiperProps' property was set as optional by using SetOptional from type-fest in ArrowSwiperWithInfoProps to enhance code robustness against potential null errors. IntervalWrapperProps was removed in in-the-news.tsx to simplify the file structure and optimize code readability. These changes aim to heighten the app's responsiveness, facilitate user navigation, and optimize the user interface.