Update import sources in React components
The import source of "MotionTypography" in the file "build-future.tsx" was removed as it was not being used leading to better performance. The import source of "motion-elements" was updated to "motion-components" in the "Sustainability.tsx" file. This change was necessary for the proper functioning of the "MotionButton" and other related components after the source refactor.