Refactor StorySlideProps and move data to OurStories component
The properties inside the 'data' object of StorySlideProps were removed and added directly to the StorySlideProps. As a response, also updated the deconstruction that was happening in the StorySlide component. With this change, each property can be passed directly to StorySlideProps, increasing reusability and ease of understanding. Lastly, moved the 'data' constant to the OurStories component for better locality.