Remove unnecessary components from page.tsx
This commit removes four previously imported components that were commented out and not utilized within the page.tsx file. The components 'OurFocusAreas', 'BuildFuture', 'Section', and 'OurCaseStudies' have been removed to clean up the codebase and maintain readability. These components were previously commented out which indicates they are not necessary for the functionality of this page.