Refactor story-slide component and update content
The story-slide component was refactored, the aspectRatio prop was removed, and typography components were wrapped in a Stack component for better layout management. Styles were adjusted, textTransform was set to "capitalize" for consistency, and a text wrapping style was added to the description. This was done for improved accessibility and the enhanced reading experience of the users. The content of a story was also updated was to better reflect the real-life success story of Ankit, spotlighting the skill development initiative's impact.