Refactor home page structure and improve styling
The home page structure has been refactored with the addition of the new OurStoriesAndKeyHighlights component for a more organised and clearer code structure.

The gradient background color previously applied to the KeyHighlights section was moved to this new parent component to ensure uniformity across both the OurStories and KeyHighlights sections.

In addition, padding abbreviations in the search-drawer component were replaced with their complete form ('pt' and 'pb') to make the code more understandable.