"Update typography and refactor homepage components
Improved the typography theme by modifying the line-height for a better reading experience. Re-constructed the structure of the home/our-stories section with the introduction of location icon and other design enhancements. The gradient background of the KeyHighlights and OurStories sections have been harmonized for a seamless visual transition. Also made revisions to the Swiper component and updated company and location in the our-stories section."