Refactor animation and style properties in StorySlide and AppBar
In StorySlide, animation properties have been factored out into a 'storySlideMotionVariant' constant for clearer animation management. `initial` and `whileInView` values for the MotionStack are also updated for more straightforward reference to the variant props. AppBar is modified to implement backward reference and handle styling through the 'sx' property to ease the manipulation of styles. This makes the code more maintainable and optimizes animation performance.