Add viewport control to footer and improve copyright icon animation
Extended the viewport in footer/index.tsx by adding 'amount' attribute, allowing better control over when animations should trigger based on the proportion of the element in view. This improves the footer animation behaviour. Additionally, the copyright.tsx file was updated to include a more lively animation for the Keyboard Arrow Up icon, contributing to a richer user experience and a more visually engaging site.