Update animation trigger in common components
Animation triggers have been modified in section-header.tsx, arrow-swiper.tsx, and arrow-slide-info.tsx within the common components. The 'amount' attribute in the viewport was changed from "all" to a numerical or "some" value to allow for better control over when animations should trigger. This was done to improve the animation behavior and enrich the user experience.