Update href in AnimatedButton component in contact-us page
The href attribute in the AnimatedButton component of the 'contact-us.tsx' page was previously empty. This has been updated to "#". With an empty string, the button could cause unexpected behavior during navigation. Using "#" as a placeholder ensures that the button, when clicked, doesn't perform any action or lead to any other page.