Update font size in copyright footer
This commit changes the font size in the copyright footer of our layout. We've replaced the previously used H6_4 font size with Body2 from our typography variants, aiming for better readability and compliance with recent design system updates. This adjustment is necessary to unify the look and feel of our application's typography across all components and to increase consistency in design.