Improve responsiveness on HeroesOfInitiatives component
A set of changes is applied to improve the layout responsiveness of the heroesOfInitiatives component. Adjustments are made on left margin, text alignment and the handling of overflow on different screen sizes. The width property of the initiative-hero-card component also has been modified to ensure a consistent and appropriate width across multiple screen sizes. This will resolve the issue of the component appearing narrower than expected on medium-sized screens.