Refactor HeroesOfInitiatives component for better responsivity
The HeroesOfInitiatives component has been refactored to enhance swiper navigation and rotation. Navigation module was imported and passed to the swiper's modules prop. The navigation object was moved to a more appropriate position in the code to improve readability. Loop functionality was removed to prevent redundant module looping. Also, a key prop was added to the stack component in InitiativeHeroCard to ensure unique keys for each set of children.