Add CompanyName prop in SectionNavigation with default value
This commit introduces a new prop 'CompanyName' in SectionNavigation component. Now, this component will be able to render the name of the company as part of the alt attribute for ADA (Americans with Disabilities Act) compliance. 'CompanyName' has a default value 'Bajaj' if not provided. Also, corrected the logo image properties by including 'alt', improving accessibility and optimizing SEO.