Adjust padding in SectionNavigation
In this commit, the padding right attribute of the SectionWrapper component in SectionNavigation has been updated. This ensures that the look and feel of the component remain consistent across all screen resolutions. The 'basePx' attribute has been utilized to standardize the padding size, thus, enhancing the component's adaptability, especially on smaller screens (resolution: xs: 0, md: 0).