Modify 'swiperSection' padding and implement 'ObjectiveSection'
The padding properties in 'scrollTriggeredSwiperSection' were simplified, changing from a multi-object structure to a single line horizontal padding ('px:0'), for enhanced code readability. A new 'ObjectiveSection' component was created to provide a uniform approach to display 'objectives' across various parts of the application. The 'ObjectiveSection' was then integrated into the 'HamaaraSapna' page. Additional padding adjustments were made in 'our-stories' index to align with the changes related to the new component.