Refactored AnimatedButton, created new FundsAndPartners component and made minor updates.
Refactored AnimatedButton component to allow for more flexibility and better reusability. The component now accepts an 'href' parameter, adopting the behavior of a link or button depending on usage context.

A new FundsAndPartners component was created in `src/components/ourCompanies/funds-and-partners.tsx`. This was done to better handle logic related to funds and partnerships, aiding readability and maintainability of the codebase.

Minor updates include adjustment of the Dialog theme, changing the borderRadius to 0 in the `src/styles/theme/components/dialogVariants.ts` file, allowing for uniform modal appearance across the app.

Lastly, the 'src/commonComponents/m-link.tsx' file was updated to make the 'href' property optional, enabling more flexible usage of the MLink component. The FormDialog component was also refactored into FormDialogButton for better code organization and readability.