Improve FundsAndPartners component and refactor FormDialogButton
Refactored FormDialogButton to accommodate a ReactNode prop. This amplifies the flexibility of the component and makes it more reusable. Likewise, FundsAndPartners component was refactored and moved to a new directory for cleaner code organization. This includes addition of an 'Enquire About Fundraising' form making it a more self-contained and independent component.

The modularity of the components is improved enhancing maintainability. It also ensures uniformity of modal interfaces across the app and facilitates flexible usage of MLink component.