Remove BeginningOfChange section from Charitable Trusts page
The "BeginningOfChange" section in the 'Charitable Trusts' page was commented out from the existing layout. This was done because the section's content was not found relevant or necessary at this point in our app service. We aim to maintain only relevant and useful information for our users, which justified the requirement of this change. We will continue to update and refine our content as needed.