Refactor components, simplify imports and add new component
Refactored several components including `company.tsx` to clean up imports and also improve the usage of the MUI Box and CardContent components. The changes improve the readability and maintenance of these components. Additionally, a  brand new component `how-we-help.tsx` was created to add more functionality to the app. This component highlights our initiatives for a brighter future within the Hamaara Sapna section.