Update common components and dependencies
Typo in the 'description' property of the ExpandableAccordionProps interface in the expandable-accordion.tsx file is corrected. An image source is added in the arrow-slide-default-image.tsx file. Also, several dependencies in package.json file are updated.

Correcting the typo ensures the proper description is displayed when data is passed via this prop and having an image source set by default will prevent the system from breaking in case no source link is provided. The update in dependencies ensures that the project utilizes the latest features, security patches and bug fixes provided by the packages.