Fix display issues in components and enhance UX
A space was added to the placeholder string in input-validator.tsx to ensure words are correctly separated, improving readability. In programmes.tsx, ReactNode import was added for type annotation while replacing generic SyntheticEvent. SelectProps were added to 'session' and 'volunteerCenter' fields for a better UI. These include a rendering method that adds a proposal to the dropdown when no value exists, and a material-ui down arrow icon for visual appeal. These changes aim to provide clearer instructions and better user experience.