Enhance ObjectiveSection Responsiveness
Center aligned images on smaller screens and updated the select properties behavior in the ObjectiveSection component. The changes were made to enhance the adaptability of UI on various devices. The addition of mx:auto will mark-up images the centered. The commenting out of InputFieldsProps and relocating those properties directly to the select objects will make those properties more effective.