Add new component HistoricLegacy for wardha charity trust page
Added HistoricLegacy component depicting the historical importance of Wardha in the freedom movement. This enhances the user's understanding and adoption of the Wardha charity trust. The data includes a description about Mahatma Gandhi's and Shri Jamanalal Bajaj's contribution along with an appropriate image. We believe this comprehensive info section will result in increased engagement levels on the site.