Add AwardsSwiper to the CharitableTrustAwards Page and refactor components
Implemented the AwardsSwiper component in the CharitableTrustAwards page to present charity awards in a more engaging way. Deleted unnecessary elements in the ObjectiveSection to improve readability, and expanded the SectionHeader component with a responsive 'isArrow' function for enhanced UX. This function provides a visual guide to the next section on different screen sizes. A new file, 'awards-swiper.tsx', was created to house the new swiper feature.