Add Award Swiper and refactor in components
Implemented an interactive Awards Swiper component in the CharitableTrustAwards page to improve user interface and engagement. Renamed 'ImageSlicsProps' to 'ImageSliceProps' for better clarity and consistency within the codebase. This includes the addition of award images and creation of a new 'awards-associated-with-us.tsx' file to keep the awards-related features organized and maintainable. This change provides a more engaging user interaction and helps maintain code consistency.